pyproject.toml
[tool.poetry]
name = "chatstatbot"
version = "0.1.0"
description = ""
authors = ["Bogdan <evstrat.bg@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.1"
python-telegram-bot = "^12.7"
python-dotenv = "^0.13.0"
pysocks = "^1.7.1"
requests = "^2.23.0"
chatstats = "^0.1.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
